pyproject.toml
[project]
name = "mcp-memory-server"
version = "1.0.0"
description = "Production-ready MCP server for LLM memory management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AgentMemorh", email = "agent@memorh.dev"}
]
keywords = ["mcp", "llm", "memory", "ai", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastmcp>=2.14.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "slowapi>=0.1.9",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "ruff>=0.1.0",
    "mypy>=1.6",
    "types-PyYAML>=6.0",
    "httpx>=0.25",
]

[project.scripts]
memory-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prometheus_client.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
